refactor(repl): migrate fabric-generation commands into FabricGenCommandSet
Move the 15 fabric-generation commands (gen_config_mem, gen_switch_matrix,
gen_tile, gen_all_tile, gen_fabric, gen_geometry, gen_bitStream_spec,
gen_top_wrapper, run_fab, run_FABulous_fabric, gen_model_npnr, gen_io_tiles,
gen_io_fabric, gen_io_pin_config, generate_custom_tile_config) into
FabricGenCommandSet.

generate_custom_tile_config takes the positive --switch-matrix flag (enabled by
default) rather than the original --no-switch-matrix. A boolean Option literally
named "--no-switch-matrix" is inferred as an argparse BooleanOptionalAction,
which Python 3.13+ rejects because it reserves the --no- prefix for the
auto-generated negation form. argparse still generates --no-switch-matrix as the
negation, so that flag keeps working and keeps its "skip the switch matrix"
meaning.

stack-info: PR: https://github.com/FPGA-Research/FABulous/pull/893, branch: KelvinChung2000/stack/7
